Sensing overhaul
This commit includes several co-related changed regarding the sensing API.
- Builtin sensing identifier extension
  + More metrics: cpu core count per disk, per interface readings
  + Exponential average of metrics (ema)
- Plugin ABI change
  + Plugins expose data buffer where cached sensing data is read from (no designated read call)
  + Plugins can report more rigorous error messages (message call)
  + Sid bit semantic changes
    * sid == 0 no more indicates error
  + can signal sensing value type (with MSB of sid)
- Boolean sensing value type is supported
- Repeated errors from sensors during refreshing are supressed
- Drops support for plugin hotswap completely
  + It require 'dynamic binding on runtime' + 'unresolved errors on runtime'
  + pretty nasty to handle, would drag the performance
  + restarting the system does not cost much (occurs hardly, happens instantly)
- plugins/debug.toml refreshes only when updated
- Expr.rs, toml.rs now uses f64 for number types
  + largest integer without precision loss is 2^24 for f32, 2^53 for f64
  + 2^24 is too small to carry memory size in bytes. (~16 MiB)

Bug Fixes
- Change in parameter value in list.toml does not properly updates SpriteDecls parameter.
  + Old impl of PartialEq for Params only checked for keys
- Some typos

+++ app/src/parser/toml.rs
@@ -66,7 +66,7 @@ impl From<WithPos<LexError>> for WithPos<ParseError<'static>> {
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Token<'src> {
     Identifier(&'src str),
-    Number(f32),
+    Number(f64),
     Boolean(bool),
     String(&'src str),
     Equals,
@@ -242,7 +242,7 @@ impl<'src> Lexer<'src> {
                 }
                 let s = &self.input[start..end];
                 pos.span = Span { start, end };
-                let num: f32 = s.parse().map_err(|_| WithPos { val: LexError::InvalidNumber, pos })?;
+                let num: f64 = s.parse().map_err(|_| WithPos { val: LexError::InvalidNumber, pos })?;
                 Ok(WithPos { val: Token::Number(num), pos })
             }
             _ => Err(WithPos { val: LexError::UnexpectedChar, pos }),
@@ -256,7 +256,7 @@ impl<'src> Lexer<'src> {
 #[derive(Debug, Clone)]
 pub enum Value<'src> {
     String(&'src str),
-    Number(f32),
+    Number(f64),
     Boolean(bool),
     Array(Vec<WithPos<Value<'src>>>),
     Table(Table<'src>),
@@ -315,7 +315,7 @@ macro_rules! extract_value_impl {
 }
 
 extract_value_impl!{&'src str, "string", String}
-extract_value_impl!{f32, "number", Number}
+extract_value_impl!{f64, "number", Number}
 extract_value_impl!{bool, "boolean", Boolean}
 extract_value_impl!{Vec<WithPos<Value<'src>>>, "array", Array}
 extract_value_impl!{Table<'src>, "table", Table}
